Use float instead of int with the same 1/8th scale as the rest of the gui

+++ gui.h
@@ -81,7 +81,7 @@ struct queue {
     void addRectangle(float x, float y, float w, float h, float r, uint32_t color);
     void addTriangle(int x, int y, int w, int h, int flags, uint32_t color);
     void addText(int x, int y, int align, const u::string &contents, uint32_t color);
-    void addImage(int x, int y, int w, int h, const u::string &path);
+    void addImage(float x, float y, float w, float h, const u::string &path);
 private:
     u::stack<command, kCommandQueueSize> m_commands;
 };
@@ -140,7 +140,7 @@ void drawRectangle(float x, float y, float w, float h, uint32_t color);
 void drawRectangle(float x, float y, float w, float h, float r, uint32_t color);
 void drawText(int x, int y, int align, const u::string &contents, uint32_t color);
 void drawTriangle(int x, int y, int w, int h, int flags, uint32_t color);
-void drawImage(int x, int y, int w, int h, const u::string &path);
+void drawImage(float x, float y, float w, float h, const u::string &path);
 
 const queue &commands();
